added two actionListener classes and multiplayer class which initializes multiplayer game
