+++ src/main/java/chess/actions/ActionOpen.java
@@ -9,6 +9,7 @@ import com.thoughtworks.xstream.XStream;
 import javax.swing.*;
 import javax.swing.filechooser.FileNameExtensionFilter;
 import javax.swing.filechooser.FileSystemView;
+import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
@@ -66,7 +67,7 @@ public class ActionOpen implements ActionListener {
     }
 
     public void resumeSavedGame(GameAttributes savedGameAttributes) { // resumes saved game
-        JFrame gameParametersWindow = menu.createGameParametersWindow(2, 2);
+        JFrame gameParametersWindow = createGameParametersWindow(2, 2);
 
         // window objects
         JLabel enterNameLabel = new JLabel("  Your name:");
@@ -86,5 +87,16 @@ public class ActionOpen implements ActionListener {
             gameParametersWindow.dispose();
             menu.getMenuWindow().dispose();
         });
+    }
+
+    public JFrame createGameParametersWindow(int rows, int columns) {
+        JFrame gameParametersWindow = new JFrame("Game Parameters");
+
+        // window parameters
+        gameParametersWindow.setSize(new Dimension(400, 100));
+        gameParametersWindow.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
+        gameParametersWindow.setVisible(true);
+        gameParametersWindow.setLayout(new GridLayout(rows, columns));
+        return gameParametersWindow;
     }
 }
