+++ src/main/java/chess/mechanics/Board.java
@@ -292,6 +292,8 @@ public class Board extends JPanel implements MouseListener {
             if (isMoved || isMoved2) {
                 repaint();
                 System.out.println("One of the components was moved");
+                switchTurn();
+                setBoardAltered(true); // sets board status as altered
                 clearMoves();
                 Cell.Colour playerColour = whiteTurn ? Cell.Colour.white : Cell.Colour.black;
                 Move move = new Move(cells[prevY][prevX].getPosition(), clicked.getPosition());
@@ -319,8 +321,6 @@ public class Board extends JPanel implements MouseListener {
                     }
 
                 }
-                switchTurn();
-                setBoardAltered(true); // sets board status as altered
             }
         }
         repaint();
